src/requirements/base.txt
asgiref==3.6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
backports-zoneinfo==0.2.1 ; python_full_version >= "3.8.1" and python_version < "3.9"
certifi==2022.12.7 ; python_full_version >= "3.8.1" and python_version < "4"
cfgv==3.3.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
charset-normalizer==3.1.0 ; python_full_version >= "3.8.1" and python_version < "4"
coreapi==2.3.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
coreschema==0.0.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
distlib==0.3.6 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
django-environ==0.10.0 ; python_full_version >= "3.8.1" and python_version < "4"
django==4.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
djangorestframework==3.14.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
drf-yasg==1.21.5 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
filelock==3.11.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
flake8==6.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
identify==2.5.22 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
idna==3.4 ; python_full_version >= "3.8.1" and python_version < "4"
inflection==0.5.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
itypes==1.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
jinja2==3.1.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
markupsafe==2.1.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
mccabe==0.7.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
nodeenv==1.7.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
packaging==23.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pillow==9.5.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
platformdirs==3.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pre-commit==3.2.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
psycopg2==2.9.6 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pycodestyle==2.10.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyflakes==3.0.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytz==2023.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyyaml==6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
requests==2.28.2 ; python_full_version >= "3.8.1" and python_version < "4"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_version < "3.11" and python_full_version >= "3.8.1"
ruamel-yaml==0.17.21 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
setuptools==67.6.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
sqlparse==0.4.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
tzdata==2023.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
urllib3==1.26.15 ; python_full_version >= "3.8.1" and python_version < "4"
virtualenv==20.21.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
